utils pagination&oder

+++ viewModels/paginator.go
@@ -1,7 +1,7 @@
 package viewModels
 
 type Paginator struct {
-	TotalRecord int         `json:"total_record"`
+	TotalRecord int64         `json:"total_record"`
 	Records     interface{} `json:"records"`
 	Limit       int         `json:"limit"`
 	Page        int         `json:"page"`
